requirements.txt
fastapi
uvicorn
python-multipart
jinja2
supabase
scikit-learn
umap-learn
hdbscan
pandas
numpy
requests
python-dotenv
tenacity
httpx
pyjwt
